ml-service/requirements.txt
torch==2.3.1
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
transformers==4.41.2
Pillow==10.3.0
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
scikit-learn==1.5.0
numpy==1.26.4
